requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1
sentencepiece==0.1.99
